requirements.txt
anthropic>=0.39.0
pyyaml>=6.0.1
python-dotenv>=1.0.0
pytest>=7.4.0
pytest-cov>=4.1.0
black>=23.0.0
flake8>=6.0.0
mypy>=1.5.0

